requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy[asyncio]
aiosqlite
asyncpg
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
